% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juicer2mat.R
\name{juicer2mat}
\alias{juicer2mat}
\title{calculate the reproducibility of HiChIP data in .hic format and display all inter-chromosome and intra-chromosome in data frame and heat map}
\usage{
juicer2mat(dic, chromfile_path, pat1 = "_r1", pat2 = "_r2", len,
  maxrange = 5e+06, resolution = 5e+05, mapHeat = TRUE)
}
\arguments{
\item{dic}{path of the contant maps exist}

\item{chromfile_path}{a file that contain chromosome length information that is used for calculating bin size}

\item{pat1}{pattern of files for chromosome 1. Default is "_r1".}

\item{pat2}{pattern of files for chromosome 2. Default is "_r2".}

\item{len}{amount of chromosome that will be compared about, which count from chr1}

\item{maxrange}{An integer indicating the maximum distance of interaction that is considered. Default is 5000000.}

\item{resolution}{An integer indicating the resolution level of the HiChIP matrix. Default is 500000.}

\item{mapHeat}{If need to display in heat map or not. Default is TRUE}
}
\value{
matrix giving number of all inter-chromosome and intra-chromosome comparasion
}
\description{
calculate the reproducibility of HiChIP data in .hic format and display all inter-chromosome and intra-chromosome in data frame and heat map
}
